// apps/client/src/data/wearable-quality.ts
export type QualityTier =
  | 'broken'
  | 'budget'
  | 'average'
  | 'excellent'
  | 'flawless';
// ...
export const DEFAULT_QUALITY_TIER: QualityTier = 'average';